refactor: create dedicated endpoint for shrink wrap first message
